+++ src/main/java/roomescape/service/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.service.dto;
 
 import java.time.LocalDate;
 import roomescape.domain.Person;
@@ -7,10 +7,6 @@ import roomescape.domain.Reservation;
 public record ReservationRequest(String name, LocalDate date, long timeId) {
 
     public Reservation toReservation() {
-        return new Reservation(toPerson(), date);
-    }
-
-    private Person toPerson() {
-        return new Person(name);
+        return new Reservation(new Person(name), date);
     }
 }
