+++ src/main/java/roomescape/service/dto/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
